feat: 다이어리 라우트 설정
