+++ mongeul/src/app/(is-login)/diary/@detail/page.tsx
@@ -1,0 +1,3 @@
+export default function Page() {
+  return <div>다이어리 - 기본</div>;
+}
