+++ mongeul/src/app/(is-login)/diary/layout.tsx
@@ -1,0 +1,12 @@
+import ParellelLayout from "@/components/layout/ParellelLayout";
+import { ReactNode } from "react";
+
+export default function Layout({
+  children,
+  detail,
+}: {
+  children: ReactNode;
+  detail: ReactNode;
+}) {
+  return <ParellelLayout detail={detail}>{children}</ParellelLayout>;
+}
